docs: update README with docs link and getting started
Add a link to the full documentation at the top of the README to help users find detailed usage information. Introduce a new "Getting started" section with instructions for installing the package and creating a new .NET web project, making it easier for new users to begin using the library.